Refactor code into discord.py Cogs
