+++ .gitignore
@@ -1,3 +1,3 @@
-*neekokeys.py
+keys.py
 __pycache__*
